[CapturingServer] fix bug relating to unzip

+++ src/main/java/com/capturingserver/services/impl/ImageServiceImpl.java
@@ -55,38 +55,41 @@ public class ImageServiceImpl implements ImageService {
     public void handle3dZipFolder(List<String> modelPaths) {
         modelPaths.forEach(modelPath -> CompletableFuture
                 .supplyAsync(() -> {
+                            String localModelFolderPath = ROOT_FOLDER_LOCATION + modelPath;
+                            String remoteModelFolderPath = REMOTE_ROOT_FOLDER_LOCATION + modelPath;
+                            String remoteModelFileNamePath = remoteModelFolderPath + Constants.FRONT_SLASH + modelPath;
+                            String capturedImageFolderPath = localModelFolderPath + Constants.BACKSLASH + Constants.CAPTURED_IMAGES_FOLDER;
                             try {
-                                String localModelFolderPath = ROOT_FOLDER_LOCATION + modelPath;
-                                String remoteModelFolderPath = REMOTE_ROOT_FOLDER_LOCATION + modelPath;
-                                String remoteModelFileNamePath = remoteModelFolderPath + Constants.FRONT_SLASH + modelPath;
-                                String capturedImageFolderPath = localModelFolderPath + Constants.BACKSLASH + "capturedImages";
+                                // Clean
                                 FileUtils.deleteDirectory(new File(localModelFolderPath));
                                 Path localModelFolderZipPath = Paths.get(localModelFolderPath + FilenameUtils.EXTENSION_SEPARATOR + Constants.ZIP_EXTENSION);
                                 if(Files.exists(localModelFolderZipPath)){
                                     Files.delete(localModelFolderZipPath);
                                 }
+
                                 scpService.transferFileFromSftpServer(remoteModelFileNamePath, ROOT_FOLDER_LOCATION, Constants.ZIP_EXTENSION, true);
-                                unzip3dFolder(localModelFolderPath + FilenameUtils.EXTENSION_SEPARATOR + Constants.ZIP_EXTENSION);
+                                unzip3dFolder(localModelFolderPath + FilenameUtils.EXTENSION_SEPARATOR + Constants.ZIP_EXTENSION, false);
                                 ProcessBuilder builder = new ProcessBuilder().redirectErrorStream(true);
                                 String command = SERVER_3DS_COMMAND + " -mxsString root:\"" + localModelFolderPath.replace("\\","\\\\")
                                         + "\" -mxsString baseFolder:\"" + capturedImageFolderPath.replace("\\","\\\\") + "\" -listenerLog \"test.log\"";
+                                System.err.println("Command: " + command);
                                 if (isWindows()) {
                                     builder.command("cmd.exe", "/c", command);
                                 }
                                 Process process = builder.start();
 
                                 int exitCode = process.waitFor();
-                                System.err.println("result: " + (exitCode != 0));
+                                System.err.println("======Capture has done with status: " + exitCode);
                                 if (exitCode != 0) {
-                                    System.err.println("result: " + (exitCode != -130));
                                     if(exitCode != -130) {
                                         throw new InterruptedException("Cannot capture 3d model ! Exit code=" + exitCode);
                                     }
                                 }
                                 String zipPath = capturedImageFolderPath + FilenameUtils.EXTENSION_SEPARATOR + Constants.ZIP_EXTENSION;
+                                //Zip capturedImages folder then send to remote Server by SFTP
                                 zipDirectory(capturedImageFolderPath, zipPath);
                                 scpService.transferToSftpServer(zipPath, remoteModelFolderPath, false);
-                                FileUtils.deleteDirectory(new File(localModelFolderPath));
+
                                 MultiValueMap<String, Object> requestBody = new LinkedMultiValueMap<>();
                                 requestBody.add(Constants.DATA, modelPath);
                                 HttpEntity<MultiValueMap<String, Object>> requestEntity = new HttpEntity<>(requestBody);
@@ -101,6 +104,12 @@ public class ImageServiceImpl implements ImageService {
                                 requestBody.add(Constants.ERROR, e.getMessage());
                                 HttpEntity<MultiValueMap<String, Object>> requestEntity = new HttpEntity<>(requestBody);
                                 new RestTemplate().postForObject(SEARCH_ENGINE_SERVER_URL + CommonRESTRegistry.CAPTURING_IMAGE_STATUS_RESPONSE, requestEntity, String.class);
+                            } finally {
+                                try {
+                                    FileUtils.deleteDirectory(new File(localModelFolderPath));
+                                } catch (IOException e) {
+                                    logger.error("Could not delete folder after capturing: " + e.getMessage());
+                                }
                             }
                             return null;
                         }
@@ -108,42 +117,43 @@ public class ImageServiceImpl implements ImageService {
     }
 
 
-    private void unzip3dFolder(String modelPath) throws IOException {
-        ZipInputStream zipInputStream = new ZipInputStream(new FileInputStream(modelPath));
-        String extractedModelDirectory = modelPath.substring(0, modelPath.lastIndexOf('.'));
-        Files.createDirectories(Paths.get(extractedModelDirectory));
-        ZipEntry entry;
-        while ((entry = zipInputStream.getNextEntry()) != null) {
-            if (entry.isDirectory()) {
-                Files.createDirectories(Paths.get(extractedModelDirectory + Constants.BACKSLASH + entry.getName()));
-            } else {
-                String entryName = entry.getName();
-                if (Paths.get(entryName).getParent() != null) {
-                    createFolderRecursively(extractedModelDirectory + Constants.BACKSLASH + Paths.get(entryName).getParent());
-                }
-                String uncompressedFilename = extractedModelDirectory + Constants.BACKSLASH + entryName;
-                uncompressedFilename = StringUtils.replace(uncompressedFilename, Constants.BLANK_SPACE, Constants.UNDERSCORE);
-                Files.createFile(Paths.get(uncompressedFilename));
-                try(FileOutputStream fos = new FileOutputStream(uncompressedFilename)) {
-                    int bufferSize = 1024;
-                    try (BufferedOutputStream bos = new BufferedOutputStream(fos, bufferSize)) {
-                        int count;
-                        byte[] buffer = new byte[1024];
-                        while ((count = zipInputStream.read(buffer, 0, bufferSize)) != -1) {
-                            bos.write(buffer, 0, count);
+    private void unzip3dFolder(String modelPath, boolean keepFileAfterUnzip) throws IOException {
+        try(ZipInputStream zipInputStream = new ZipInputStream(new FileInputStream(modelPath))) {
+            String extractedModelDirectory = modelPath.substring(0, modelPath.lastIndexOf('.'));
+            Files.createDirectories(Paths.get(extractedModelDirectory));
+            ZipEntry entry;
+            while ((entry = zipInputStream.getNextEntry()) != null) {
+                if (entry.isDirectory()) {
+                    Files.createDirectories(Paths.get(extractedModelDirectory + Constants.BACKSLASH + entry.getName()));
+                } else {
+                    String entryName = StringUtils.replace(entry.getName(), Constants.BLANK_SPACE, Constants.UNDERSCORE);
+                    if (Paths.get(entryName).getParent() != null) {
+                        createFolderRecursively(extractedModelDirectory + Constants.BACKSLASH + Paths.get(entryName).getParent());
+                    }
+                    String uncompressedFilename = extractedModelDirectory + Constants.BACKSLASH + entryName;
+                    Files.createFile(Paths.get(uncompressedFilename));
+                    try(FileOutputStream fos = new FileOutputStream(uncompressedFilename)) {
+                        int bufferSize = 1024;
+                        try (BufferedOutputStream bos = new BufferedOutputStream(fos, bufferSize)) {
+                            int count;
+                            byte[] buffer = new byte[1024];
+                            while ((count = zipInputStream.read(buffer, 0, bufferSize)) != -1) {
+                                bos.write(buffer, 0, count);
+                            }
+                            bos.flush();
                         }
-                        bos.flush();
                     }
                 }
             }
         }
-        zipInputStream.close();
-        Files.delete(Paths.get(modelPath));
+
+        if(!keepFileAfterUnzip) {
+            Files.delete(Paths.get(modelPath));
+        }
     }
 
 
     public static void zipDirectory(String sourceDirectoryPath, String zipPath) throws IOException {
-
         Path zipFilePath = Files.createFile(Paths.get(zipPath));
         try(OutputStream outputStream = Files.newOutputStream(zipFilePath)) {
             try (ZipOutputStream zipOutputStream = new ZipOutputStream(outputStream)) {
